changed insert methods
now the methods no longer throw any sqllite excaption but rather return true or false whether or not it succesfully completed the task.
The info_timestamp table now contains a successfull column to indicate whether or not the operation was executed successfully.
The needs_updated method usese the successful as an extra filter